Logic to check user status and paragraph suppression
https://eaflood.atlassian.net/browse/BNGP-3724

For the allocations journey, the view for checking a users Defra account displays two paragraphs of text currently regardless of how questions are answered. This change enhances allocation journey logic to take account of whether the user is not acting on behalf of a client and is not the landowner/leaseholder (non-relevant person) and suppresses the second paragraph if so.