Bugfix (Check Your Answers link disabled)
After completing the last task, the Check Your Answers link should be enabled when we return to the task list. However, we found that this was not happening right away -- the link would only become enabled after refreshing. This was because we were creating a shallow copy of the Check Your Answers section, so deleting the href property from the copy caused the href property to be deleted from the original, hence it not being enabled when returning to the task list. We therefore fix this by deep cloning the section before adding it to the task list.