feat: Add link to register biodiversity gain sites based on combined case journey support
This commit adds a link to register biodiversity gain sites on the manage-biodiversity-gains page. The link is conditionally rendered based on the value of the ENABLE_ROUTE_SUPPORT_FOR_COMBINED_CASE_JOURNEY environment variable. If the variable is set to "Y", the link will direct the user to the start register gain site page. If the variable is set to "N", the link will direct the user to the biodiversity gain sites page.