Prepare for merge
In anticipation of merge conflicts in `task-list-generator.js` caused by the additional `isLocked` functionality , we rename it to `task-list-generator-v5.js`. We will then merge in the master branch, giving us two versions of the task list generator working side-by-side, and we can then convert the combined case list to the new version.