Fix `application.spec.js` tests
A few fixes are needed in `application.spec.js`. We first remove the `.only` which prevents all the tests from running. Then we fix some unit tests as they depend on behaviour which was changed in a previous PR (they expect `optional` to be set in the file object when this is no longer done). We remove geospatial-specific tests. Finally, we slightly amend some test descriptions to remove references to geospatial files.