Fix "Biodiversity gain site number" no change issue
Going to change "Biodiversity gain site number" but clicking Continue without a change results in parts of the application being reset and the user being returned to the task list. The routing logic was there but being executed in the wrong order -- the "no change" check was done _after_ parts of the application were reset. We fix the issue by doing the check first.