BNGP-5362: Fix incorrect flow logic when changing from Check Your Answers (#840)
* Refactor common code with updated "no change" logic

* Use `CHECK_AND_SUBMIT_JOURNEY_ROUTE` key to track if user started journey from Check And Submit

Elsewhere we have used `getValidReferrerUrl()` to determine whether a user came from a Check And Submit page and redirect them accordingly. We cannot do this on the `check-applicant-information` page as some journeys (eg. `agent-acting-for-client`) clear the referrer to ensure the user is taken through the full journey. We therefore create a new redis key `CHECK_AND_SUBMIT_JOURNEY_ROUTE` which we set when the user visits a Check And Submit page (and clear when they visit a task list page). We update the `check-applicant-information` routing logic to redirect to this url if it exists.

* Set and clear journey route key for combined case

* Remove stray `console.log()`

* Fix `legal-agreement` page