BNGP-5255 Change task list to v5 GOV UK Design System component (#813)
* Prepare `task-list-generator` for component data format

* Amend register task list to use component

* Amend allocations task list to use component

* Bugfix (Check Your Answers link disabled)

After completing the last task, the Check Your Answers link should be enabled when we return to the task list. However, we found that this was not happening right away -- the link would only become enabled after refreshing. This was because we were creating a shallow copy of the Check Your Answers section, so deleting the href property from the copy caused the href property to be deleted from the original, hence it not being enabled when returning to the task list. We therefore fix this by deep cloning the section before adding it to the task list.

* Bugfix (Wrong Check Your Answers link for allocations task list)

* Fix allocations Check Your Answers task text

* Amend credits purchase task list to use component

* Replace `lodash/cloneDeep` with `structuredClone`

Node 18+ gives us a native function to deep clone an object, so we don't need to use lodash for this

* Remove v4 task list data

While migrating the from v4 to v5, we passed around data in both the original format we used as well as the v5 format. Now that we've updated all 3 task list pages to use the component, we no longer need data in the v4 format

* Refactor task list completed check

Previously, if a task contained 1 item, we directly checked that item to see if it was completed; otherwise, we iteratively checked each item in the task. Since both paths effectively do the same thing, we refactor the code so we always iterate over a task's items.

* Fix `register-land-task-list` unit tests

* Fix `developer/tasklist` unit tests

* Fix `credits-purchase/tasklist` unit tests

* Fix `task-list-generator` unit tests

* Prepare for merge

In anticipation of merge conflicts in `task-list-generator.js` caused by the additional `isLocked` functionality , we rename it to `task-list-generator-v5.js`. We will then merge in the master branch, giving us two versions of the task list generator working side-by-side, and we can then convert the combined case list to the new version.

* Merge branch 'master' into bngp-5255-upgrade-to-v5-task-list-component

* Add v5 task list to combined case task list page

* Rename "task list" to "task list v4", and "task list v5" to "task list"

We make the v5 task list the standard task list; and treat the incoming v4 task list as the exception to be changed.

* Rename `task-list-generator` unit test files to match v4/v5 change

* Move v4 functions into `task-list-generator.js`

* Bring combined case task list over to v5 component completely

* Refactor to remove unneeded `getTaskStatus` function

* Update failing combined case tests

* Refactor to simplify how we handle status

* Fix failing unit test

* Fix failing unit tests

* Refactor to create common task list

* Remove unneeded `idPrefix` from tasklist component

* Update unit test

* Update `task-list-generator` to apply custom ids to each task and its status

* Fix unit tests

* Give `Not started yet` status to check & submit

* Remove custom styling and nested lists from `tasklist.html` template

Now that we're using the govuk tasklist component we remove the custom styling from the `tasklist.html` template. We also remove nested lists to ensure WCAG compliance.

* Fix `register-land-task-list.spec.js`

* Fix `developer/tasklist.spec.js`

* Fix `combined-case/tasklist.spec.js`

* Fix `credits-purchase/tasklist.spec.js`