Fix issue where user cannot upload new files
If the user has completed the "Add land ownership details" section, they cannot upload a new file -- selecting the "No, add another proof of land ownership file" returns them back to where they came (either the task list or Check Your Answers). This is because the `LAND_OWNERSHIP_PROOF_LIST_KEY` key used to check if this section is complete is not cleared if they decide to upload a new file, and the `nextUrl` function of this route returns the user to the task list or Check Your Answers if this is set. We therefore clear this key if the user chooses to upload another file.