Direct user back to Check Your Answers if their land ownership proof started there
Both the registration and combined case had logic in the land ownership page to direct the user back to the referrer if one exists, or back to the task list if not. The problem is that the referrer is reset if they make changes during the journey, so they would be returned to the task list regardless of where they started the journey. We replace the use of the referrer with the `CHECK_AND_SUBMIT_JOURNEY_ROUTE` key, which has the Check Your Answers url if the journey started there.