Initial `sectionIdHandler` plugin
We're going to identify which section the user came from by appending a query param to each link on the check and submit page. We will intercept requests with this param and store it, then redirect to the page without the query param (so that this is effectively invisible to the user). Finally, when the user visits check and submit, we will check for the presence of the stored param and take the user to the correct section if so.

The first stage is to intercept requests and look for the query param. We do this in an initial `sectionIdHandler` plugin, which checks for a `journey-start-section-id` param and if found, it logs it to the console and redirects to the url without the param.