Fix bug in `generate-summary-list-macro.njk` and refactor
We spotted a bug in `generate-summary-list-macro.njk`-- the id attribute for each item was defined as `id="{{item.valueId}}}"`, ie. 3 closing braces. This therefore appended an extra `}` to each id. Furthermore, not every item has an id, leading to blank `id` attributes. We replace the id attribute entirely with conditional logic that sets it or not depending on whether an id is passed. Since we're potentially applying the `generateSummaryList` macro to summaries that don't have defined data test ids, we also make `data-testid` conditional in the same way.