BNGP-5361: Fix proof of land ownership logic flow (#852)
* Fix issue where user cannot upload new files

If the user has completed the "Add land ownership details" section, they cannot upload a new file -- selecting the "No, add another proof of land ownership file" returns them back to where they came (either the task list or Check Your Answers). This is because the `LAND_OWNERSHIP_PROOF_LIST_KEY` key used to check if this section is complete is not cleared if they decide to upload a new file, and the `nextUrl` function of this route returns the user to the task list or Check Your Answers if this is set. We therefore clear this key if the user chooses to upload another file.

* Direct user back to Check Your Answers if their land ownership proof started there

Both the registration and combined case had logic in the land ownership page to direct the user back to the referrer if one exists, or back to the task list if not. The problem is that the referrer is reset if they make changes during the journey, so they would be returned to the task list regardless of where they started the journey. We replace the use of the referrer with the `CHECK_AND_SUBMIT_JOURNEY_ROUTE` key, which has the Check Your Answers url if the journey started there.