BNGP-5091: Focus on updated field when returning to check and submit page (#854)
* Add section ids to page

* Small refactor of how we pass section ids

* Initial `sectionIdHandler` plugin

We're going to identify which section the user came from by appending a query param to each link on the check and submit page. We will intercept requests with this param and store it, then redirect to the page without the query param (so that this is effectively invisible to the user). Finally, when the user visits check and submit, we will check for the presence of the stored param and take the user to the correct section if so.

The first stage is to intercept requests and look for the query param. We do this in an initial `sectionIdHandler` plugin, which checks for a `journey-start-section-id` param and if found, it logs it to the console and redirects to the url without the param.

* Refactor check and submit to add `journey-start-section-id` query param

We do the following to add the `journey-start-section-id` query param to each Change link on the combined case check and submit page:

- Amend `generate-summary-list-macro.njk` to optionally accept `sectionId` in the provided summary, which is the value of the query param added to each Change link
- Amend `applicant-information.html` and `legal-agreement-summary.html` to pass on the section id (note that this also adds the section id to the register journey; if we get to refactor that check and submit page then this stands us in good stead; otherwise, it makes no difference)
- Refactor `check-and-submit.html` to use `generateSummaryList` in place of `govukSummaryList` so we can set the query param on each link;
- Amend `get-land-check-and-submit-details.js` to pass land ownership files data through in the format used by `generateSummaryList` as well as `govukSummaryList`; although we've refactored the combined case check and submit page to use `generateSummaryList`, we haven't (yet) refactored the register journey check and submit to use this. So for now we pass the data through in both formats.

* Add `JOURNEY_START_SECTION_ID` key and use it inside `section-id-handler.js`

We create a redis key to store the value of `journey-start-section-id` and update `section-id-handler.js` to set it, and to apply it on the check and submit page

* Unit test `sectionIdHandler`

* Tweak tests slightly

* Fix bug in `generate-summary-list-macro.njk` and refactor

We spotted a bug in `generate-summary-list-macro.njk`-- the id attribute for each item was defined as `id="{{item.valueId}}}"`, ie. 3 closing braces. This therefore appended an extra `}` to each id. Furthermore, not every item has an id, leading to blank `id` attributes. We replace the id attribute entirely with conditional logic that sets it or not depending on whether an id is passed. Since we're potentially applying the `generateSummaryList` macro to summaries that don't have defined data test ids, we also make `data-testid` conditional in the same way.

* Take user to specific item, not section

On re-reading the ticket we realised we were taking the user back to the section they had come from when the ticket calls for them to return to the changed item. We update our code to do this.

* Remove section ids

Now that we take the user to the changed response rather than the section, we remove all section-related ids and constants

* Rename "section id handler" to "answer id handler"

* Move page to be monitored into `constants.js`

We're going to apply the answerIdHandler logic to the combined case task list. Ahead of this, we create an array `constants.answerIdRoutes` which we will use to store all the pages the logic should apply to.

* Clear stored answer id if user "breaks out" of journey

* Clear answer id when restoring a session to prevent unwanted behaviour

Because the answer id is stored in the session, restoring a session will restore any existing value which can have undesired behaviour (jumping to an answer if they revisit check and submit when this is no longer required). We therefore clear this value when restoring a session.

* Enable functionality for combined case task list

* Fix unit tests

* Typo

* Small formatting fix

* Use array for `JOURNEY_START_ANSWER_ID` instead of string

* Update `answer-id-handler` to fully implement call stack

* Fix unit tests

* Fix 500 error when session not initialised

---------

Co-authored-by: Shaun Ricketts <89984496+shaundefra@users.noreply.github.com>