BNGP-5355 Remove unusued geospatial code (#837)
* Stub `geoprocessing-service`

* Remove licencing section from `NOTICE` for now-unused OS data

* Remove additional references to `geoprocessing-service`

* Remove `geospatial-land-boundary` upload type support from Azure function `ProcessTrustedFile`

* Remove geoprocessing dependency from Azure functions

* Delete `geoprocessing-service` completely

* Remove remaining references to `geoprocessing-service`

* Disable geospatial routes entirely

We remove the geospatial routes from `disabled-routes-constants.js` to ensure that they are never made live, regardless of the `ENABLE_ROUTE_SUPPORT_FOR_GEOSPATIAL` env var.

* Remove files of optional routes

* Remove geospatial routes from `router.spec.js`

* Remove references to `CHOOSE_LAND_BOUNDARY_UPLOAD` and `chooseLandBoundUploadHref`

* Remove `UPLOAD_GEOSPATIAL_LAND_BOUNDARY` and associated references

Note that the `isGeosptialDisabled` variable we remove is deceptively named, as it's set to `true` if `ENABLE_ROUTE_SUPPORT_FOR_GEOSPATIAL` is `Y` -- in other words, it would have been better named `isGeospatialEnabled`

* Remove references to unused `ORDNANCE_SURVEY_API` env vars

* Remove most references to `ENABLE_ROUTE_SUPPORT_FOR_GEOSPATIAL`

* Remove geospatial function from `land-boundary-context.js`

Note that we have renamed this from `geospatial-or-land-boundary-context.js` to reflect the fact we no longer have any geospational function in it

* Remove `context.geospatial`

We realise we're no longer files of type `geospatialData` so we have no need for `geospatial` in `context`

* Remove `GEOSPATIAL_DATA` constant

* Remove `GEOSPATIAL_LEGAL_LAND_BOUNDARY_FILE_EXT` constant

* Remove remaining  `ENABLE_ROUTE_SUPPORT_FOR_GEOSPATIAL` flags

* Remove unneeded errors

* Remove overlooked `ENABLE_ROUTE_SUPPORT_FOR_GEOSPATIAL` reference

* Remove remaining references to `geospatialData`

* Remove several unneeded constants

* Delete `map.js`

* Fix rebase issue

* Revert accidental test runner changes

* Fix failing unit tests following rebase

* Fix `application.spec.js` tests

A few fixes are needed in `application.spec.js`. We first remove the `.only` which prevents all the tests from running. Then we fix some unit tests as they depend on behaviour which was changed in a previous PR (they expect `optional` to be set in the file object when this is no longer done). We remove geospatial-specific tests. Finally, we slightly amend some test descriptions to remove references to geospatial files.

* Remove remaining `geojson` references

* Remove geospatial list items from `combined-case/check-and-submit.html`

* Remove `DOWNLOAD_GEOSPATIAL_LAND_BOUNDARY` page

* Remove `GEOSPATIAL_UPLOAD_LOCATION` constant

* Remove references to `confirmGeospatialLandBoundary`

* Fix failing unit test after removing `GEOSPATIAL_UPLOAD_LOCATION`

* Remove `landBoundaryUploadType` from `land-boundary-context.js`

We have no references to `landBoundaryUploadType` anywhere else in the code so we remove it.

* Remove `CHECK_GEOSPATIAL_FILE` and `COMBINED_CASE_CHECK_GEOSPATIAL_FILE` constants

These were either overlooked previously or they slipped back in during a rebase. Either way, they're not needed so we remove them.

* Remove `hideClass(geospatial)` from `check-and-submit` pages

We've removed `geospatial` from the page context so `hideClass(geospatial)` will never result in check and submit items being hidden, so we can remove each `classes: hideClass(geospatial)` entirely.

* Entirely remove `geospatial-land-boundary` from `azure-functions`

* `package.json` and `package-lock.json` fix