BNGP-1140: Credits journey back button behaviour
https://eaflood.atlassian.net/browse/BNGP-1140

This change addresses a bug where users see previously shown error messages when using the back button to navigate to previous pages in the credits journey. This fix ensures that error pages follow the Post/Redirect/Get (PRG) pattern, preventing the browser from caching the error state and displaying it when the back button is pressed.