BNGP-5394: Fix allocation check your answers flow logic (#843)
* Create `CHECK_AND_SUBMIT_JOURNEY_ROUTE` redis key

Note that this is also introduced in the as-yet unmerged branch `bngp-5362-fix-incorrect-flow-logic-when-selecting-no-change`

* Set key value in task list and check and submit page

* Fix "Biodiversity gain site number" no change issue

Going to change "Biodiversity gain site number" but clicking Continue without a change results in parts of the application being reset and the user being returned to the task list. The routing logic was there but being executed in the wrong order -- the "no change" check was done _after_ parts of the application were reset. We fix the issue by doing the check first.

* Fix "Statutory biodiversity metric file" change issue

* Fix "Acting on behalf of a client" change issue

* Fix "landowner or leaseholder" no change issue

* Fix "landowner or leaseholder" change issue

* Unit test `landowner-or-leaseholder.js`

* Unit test `confirm-off-site-gain.js`

* Refactor `confirm-off-site-gain.js` for consistency

* Unit test "acting on behalf of a client" change issue

---------

Co-authored-by: Shaun Ricketts <89984496+shaundefra@users.noreply.github.com>