Refactor check to validate file type against file extension
Since the `file-type` module gives us the expected file extension, we can simply rework the check to confirm that the file's extension matches what it's expected to be.

NOTE: we tried saving out a .doc file from MS Word for Mac 16.90 and the resulting file was identified on uploading as `cfb` -- so further investigation may be required.